Split local and production dockerfile
The local dockerfile used via dockercompose is important for developers
and designers. It allows easy changes of design without needing to
create a development environment. This means the container needs access
to the host machine to allow for mount binds. this is not wanted in
production. here we want to restrict the access of the docker user.
in these cases the container will use normal volumes.